+++ packages/components/package.json
@@ -31,6 +31,7 @@
     "@opentiny/vue-switch": "^3.21.0",
     "@opentiny/vue-select": "^3.21.0",
     "@opentiny/vue-option": "^3.21.0",
+    "@opentiny/vue-radio-group": "^3.21.0",
     "@vueuse/core": "^13.1.0",
     "markdown-it": "^14.1.0"
   },
